register , update 페이지 작성 및 기능 구현

+++ src/main/java/baseline/version3/springboot/pageAdmin/repository/ParentPageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParentPageRepository extends JpaRepository<ParentPage, Integer> {
+public interface ParentPageRepository extends JpaRepository<ParentPage, Long> {
 }
